refactor: move find_replace to test/utils.py
